refactor(floxhub): move Floxhub into flox-core::floxhub
Move the Floxhub struct, Transform, DEFAULT_FLOXHUB_URL, FloxhubError, and
the host-resolution tests from flox-rust-sdk::flox into a new
flox_core::floxhub module so lower crates can derive FloxHub URLs without
depending on flox-rust-sdk. Repoint all call sites to the new path; the
FloxhubToken re-export and its token tests stay in flox-rust-sdk. Add url
and itertools to flox-core.
