+++ cli/flox-core/Cargo.toml
@@ -12,26 +12,28 @@ blake3.workspace = true
 crossterm.workspace = true
 derive_more.workspace = true
 fslock.workspace = true
+itertools.workspace = true
 log.workspace = true
 # should be a dev-dependency but test doesn't apply across crates
 proptest = { optional = true, workspace = true }
 # should be a dev-dependency but test doesn't apply across crates
 proptest-derive = { optional = true, workspace = true }
+regex.workspace = true
 sentry.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+shell-escape.workspace = true
 shell_gen = { path = "../shell_gen" }
 supports-color.workspace = true
 sysinfo.workspace = true
-tracing.workspace = true
-tempfile.workspace = true
-time.workspace = true
-thiserror.workspace = true
-schemars.workspace = true
-uuid.workspace = true
-regex.workspace = true
 serde_with.workspace = true
-shell-escape.workspace = true
+schemars.workspace = true
+tempfile.workspace = true
+thiserror.workspace = true
+time.workspace = true
+tracing.workspace = true
+url.workspace = true
+uuid.workspace = true
 
 [dev-dependencies]
 indoc.workspace = true
