+++ cli/flox-core/src/lib.rs
@@ -2,6 +2,7 @@ pub mod activate;
 pub mod activations;
 pub mod canonical_path;
 pub mod data;
+pub mod floxhub;
 pub mod hook_actions;
 #[cfg(feature = "proc_status")]
 pub mod proc_status;
