+++ cli/flox/src/commands/envs.rs
@@ -245,7 +245,7 @@ mod tests {
     use std::str::FromStr;
 
     use flox_core::data::environment_ref::{EnvironmentName, EnvironmentOwner};
-    use flox_rust_sdk::flox::Floxhub;
+    use flox_core::floxhub::Floxhub;
     use flox_rust_sdk::models::environment::PathPointer;
     use indoc::formatdoc;
     use pretty_assertions::assert_eq;
